Modifica estructura de tests para separar por entidad
