+++ app/test/test_unit/test_event.py
@@ -5,19 +5,6 @@ from app.models import Event
 
 User = get_user_model()
 
-class UserModelTest(TestCase):
-    def test_create_user(self):
-        user = User.objects.create_user(username="testuser", password="testpass123")
-        self.assertEqual(user.username, "testuser")
-        self.assertTrue(user.is_active)
-        self.assertFalse(user.is_staff)
-        self.assertFalse(user.is_superuser)
-        self.assertFalse(user.is_organizer)
-
-    def test_create_organizer(self):
-        organizer = User.objects.create_user(username="organizeruser", password="testpass123", is_organizer=True)
-        self.assertTrue(organizer.is_organizer)
-
 class EventModelTest(TestCase):
     def test_create_event(self):
         organizer = User.objects.create_user(username="eventorganizer", password="pass")
